move SourceType to /domain
